requirements.txt
koi-net>=1.0.0b7
rid-lib>=3.2.3
rich
fastapi
uvicorn
python-dotenv
requests